requirements.txt
pyyaml>=6.0
requests>=2.31
yfinance>=0.2.40